Split all service interfaces into their own perspective files
