+++ CopeID.API/Services/Contributors/ContributorService.cs
@@ -4,8 +4,6 @@ using CopeID.Models.Contributors;
 
 namespace CopeID.API.Services.Contributors
 {
-    public interface IContributorService : IBaseEntityService<Contributor, ContributorQueryModel> { }
-
     public class ContributorService : BaseEntityService<Contributor, ContributorQueryModel>, IContributorService
     {
         public ContributorService(CopeIdDbContext context) : base(context) { }
