+++ CopeID.API/Services/Contributors/IContributorService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Contributors;
+using CopeID.Models.Contributors;
+
+namespace CopeID.API.Services.Contributors
+{
+    public interface IContributorService : IBaseEntityService<Contributor, ContributorQueryModel> { }
+}
