+++ CopeID.API/Services/Definitions/DefinitionService.cs
@@ -4,8 +4,6 @@ using CopeID.Models.Definitions;
 
 namespace CopeID.API.Services.Definitions
 {
-    public interface IDefinitionService : IBaseEntityService<Definition, DefinitionQueryModel> { }
-
     public class DefinitionService : BaseEntityService<Definition, DefinitionQueryModel>, IDefinitionService
     {
         public DefinitionService(CopeIdDbContext context) : base(context) { }
