+++ CopeID.API/Services/Definitions/IDefinitionService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Definitions;
+using CopeID.Models.Definitions;
+
+namespace CopeID.API.Services.Definitions
+{
+    public interface IDefinitionService : IBaseEntityService<Definition, DefinitionQueryModel> { }
+}
