+++ CopeID.API/Services/Filters/FilterService.cs
@@ -4,8 +4,6 @@ using CopeID.Models.Filters;
 
 namespace CopeID.API.Services.Filters
 {
-    public interface IFilterService : IBaseEntityService<Filter, FilterQueryModel> { }
-
     public class FilterService : BaseEntityService<Filter, FilterQueryModel>, IFilterService
     {
         public FilterService(CopeIdDbContext context) : base(context) { }
