+++ CopeID.API/Services/Filters/IFilterService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Filters;
+using CopeID.Models.Filters;
+
+namespace CopeID.API.Services.Filters
+{
+    public interface IFilterService : IBaseEntityService<Filter, FilterQueryModel> { }
+}
