+++ CopeID.API/Services/Genuses/IGenusService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Genuses;
+using CopeID.Models.Genuses;
+
+namespace CopeID.API.Services.Genuses
+{
+    public interface IGenusService : IBaseEntityService<Genus, GenusQueryModel> { }
+}
