+++ CopeID.API/Services/Photographs/IPhotographService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Photographs;
+using CopeID.Models.Photographs;
+
+namespace CopeID.API.Services.Photographs
+{
+    public interface IPhotographService : IBaseEntityService<Photograph, PhotographQueryModel> { }
+}
