+++ CopeID.API/Services/References/IReferenceService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.References;
+using CopeID.Models.References;
+
+namespace CopeID.API.Services.References
+{
+    public interface IReferenceService : IBaseEntityService<Reference, ReferenceQueryModel> { }
+}
