+++ CopeID.API/Services/References/ReferenceService.cs
@@ -4,8 +4,6 @@ using CopeID.Models.References;
 
 namespace CopeID.API.Services.References
 {
-    public interface IReferenceService : IBaseEntityService<Reference, ReferenceQueryModel> { }
-
     public class ReferenceService : BaseEntityService<Reference, ReferenceQueryModel>, IReferenceService
     {
         public ReferenceService(CopeIdDbContext context) : base(context) { }
