+++ CopeID.API/Services/Specimens/ISpecimenService.cs
@@ -1,0 +1,7 @@
+﻿using CopeID.API.QueryModels.Specimens;
+using CopeID.Models.Specimens;
+
+namespace CopeID.API.Services.Specimens
+{
+    public interface ISpecimenService : IBaseEntityService<Specimen, SpecimenQueryModel> { }
+}
