+++ CopeID.API/Services/Specimens/SpecimenService.cs
@@ -4,8 +4,6 @@ using CopeID.Models.Specimens;
 
 namespace CopeID.API.Services.Specimens
 {
-    public interface ISpecimenService : IBaseEntityService<Specimen, SpecimenQueryModel> { }
-
     public class SpecimenService : BaseEntityService<Specimen, SpecimenQueryModel>, ISpecimenService
     {
         public SpecimenService(CopeIdDbContext context) : base(context) { }
